separacion del prompt para mejor rendimiento
